Refactor: move Fetcher and Cache out of repository.py
