+++ cvmfs/repository.py
@@ -5,16 +5,11 @@ Created by René Meusel
 This file is part of the CernVM File System auxiliary tools.
 """
 
-import abc
 import os
-import tempfile
-import requests
 import collections
 from datetime import datetime
 import dateutil.parser
 from dateutil.tz import tzutc
-import shutil
-import zlib
 
 import _common
 import cvmfs
@@ -23,6 +18,7 @@ from catalog import Catalog
 from history import History
 from whitelist import Whitelist
 from certificate import Certificate
+from fetcher import RemoteFetcher, LocalFetcher
 from _exceptions import *
 
 
@@ -138,248 +134,6 @@ class CatalogTreeIterator(object):
         wrapper = self.catalog_stack.pop()
         return wrapper.get_catalog()
 
-
-class Cache(object):
-    """ Abstract base class for a caching strategy """
-
-    """ Try to get an object from the cache
-        :file_name  name of the object to be retrieved
-        :return     a file object of the cached object or None if not found
-    """
-    @abc.abstractmethod
-    def get(self, file_name):
-        pass
-
-    """ Open a transaction to accomodate a new object in the cache
-        :file_name  name of the object to be stored in the cache
-        :return     a writable file object to a temporary storage location
-    """
-    @abc.abstractmethod
-    def transaction(self, file_name):
-        pass
-
-    """ Commit a filled file object obtained via transaction() into the cache
-        :resource   a file object obtained by transaction() and filled with data
-        :return     a file object to the committed object
-    """
-    @abc.abstractmethod
-    def commit(self, resource):
-        pass
-
-
-class DummyCache(Cache):
-    """ A dummy cache uses temporary storage without actual cache logic """
-
-    def get(self, file_name):
-        return None
-
-    def transaction(self, file_name):
-        return tempfile.NamedTemporaryFile("w+b")
-
-    def commit(self, resource):
-        resource.seek(0)
-        return resource
-
-
-class DiskCache(Cache):
-    """ Maintains a fully functional and reusable disk cache """
-
-    class TransactionFile(file):
-        """ Wrapper around a writable file. The actual file will be renamed
-        to a different location once it is closed
-        """
-
-        def __init__(self, name, tmp_dir):
-            self.__final_destination_path = name
-            temp_path = tempfile.mktemp(dir=tmp_dir, prefix='tmp.')
-            super(DiskCache.TransactionFile, self).__init__(temp_path, 'w+b')
-
-        def __del__(self):
-            if not self.closed:
-                self.close()
-
-        def commit(self):
-            super(DiskCache.TransactionFile, self).close()
-            os.rename(self.name, self.__final_destination_path)
-            return open(self.__final_destination_path, "rb")
-
-    def __init__(self, cache_dir):
-        if not os.path.exists(cache_dir):
-            cache_dir = tempfile.mkdtemp(dir='/tmp', prefix='cache.')
-        self._cache_dir = cache_dir
-        self._create_cache_structure()
-        self._cleanup_metadata()
-
-    def _cleanup_metadata(self):
-        metadata_file_list = [
-            os.path.join(self._cache_dir, _common._MANIFEST_NAME),
-            os.path.join(self._cache_dir, _common._LAST_REPLICATION_NAME),
-            os.path.join(self._cache_dir, _common._REPLICATING_NAME),
-            os.path.join(self._cache_dir, _common._WHITELIST_NAME)
-        ]
-        for metadata_file in metadata_file_list:
-            try:
-                os.remove(metadata_file)
-            except OSError:
-                pass
-
-    def _create_dir(self, path):
-        cache_full_path = os.path.join(self._cache_dir, path)
-        if not os.path.exists(cache_full_path):
-            os.mkdir(cache_full_path, 0755)
-
-    def _create_cache_structure(self):
-        self._create_dir('data')
-        for i in range(0x00, 0xff + 1):
-            new_folder = '{0:#0{1}x}'.format(i, 4)[2:]
-            self._create_dir(os.path.join('data', new_folder))
-        self._create_dir(os.path.join('data', 'txn'))
-
-    def get_transaction_dir(self):
-        return os.path.join(self._cache_dir, 'data', 'txn')
-
-    def get_cache_path(self):
-        return str(self._cache_dir)
-
-    def transaction(self, file_name):
-        full_path = os.path.join(self._cache_dir, file_name)
-        tmp_dir = self.get_transaction_dir()
-        return DiskCache.TransactionFile(full_path, tmp_dir)
-
-    def commit(self, resource):
-        return resource.commit()
-
-    def get(self, file_name):
-        full_path = os.path.join(self._cache_dir, file_name)
-        if os.path.exists(full_path):
-            try:
-                # if the file has been removed by now the open method
-                # throws an exception
-                return open(full_path, 'rb')
-            except IOError, e:
-                raise FileNotFoundInRepository(full_path)
-        return None
-
-
-class Fetcher(object):
-    """ Abstract wrapper around a Fetcher """
-
-    __metadata__ = abc.ABCMeta
-
-    def __init__(self, source, cache_dir = None):
-        self.__cache = DiskCache(cache_dir) if cache_dir else DummyCache()
-        self.source = source
-
-    def _make_file_uri(self, file_name):
-        return os.path.join(self.source, file_name)
-
-    def get_cache_path(self):
-        if self.__cache:
-            return self.__cache.get_cache_path()
-
-    def retrieve_file(self, file_name):
-        """
-        Method to retrieve a file from the cache if exists, or from
-        the repository if it doesn't. In case it has to be retrieved from
-        the repository it will also be decompressed before being stored in
-        the cache
-        :param file_name: name of the file in the repository
-        :return: a file read-only file object that represents the cached file
-        """
-        return self._retrieve(file_name, self._retrieve_file)
-
-    def retrieve_raw_file(self, file_name):
-        """
-        Method to retrieve a file from the cache if exists, or from
-        the repository if it doesn't. In case it has to be retrieved from
-        the repository it won't be decompressed
-        :param file_name: name of the file in the repository
-        :return: a file read-only file object that represents the cached file
-        """
-        return self._retrieve(file_name, self._retrieve_raw_file)
-
-    def _retrieve(self, file_name, retrieve_fn):
-        cached_file_ro = self.__cache.get(file_name)
-        if cached_file_ro:
-            return cached_file_ro
-
-        cached_file_rw = self.__cache.transaction(file_name)
-        retrieve_fn(file_name, cached_file_rw)
-        return self.__cache.commit(cached_file_rw)
-
-
-    @abc.abstractmethod
-    def _retrieve_file(self, file_name, cached_file):
-        """ Abstract method to retrieve a file from the repository """
-        pass
-
-    @abc.abstractmethod
-    def _retrieve_raw_file(self, file_name, cached_file):
-        """ Abstract method to retrieve a raw file from the repository """
-        pass
-
-
-class LocalFetcher(Fetcher):
-    """ Retrieves files only from the local cache """
-
-    def __init__(self, local_repo, cache_dir = None):
-        super(LocalFetcher, self).__init__(local_repo, cache_dir)
-
-    def _retrieve_file(self, file_name, cached_file):
-        full_path = self._make_file_uri(file_name)
-        if os.path.exists(full_path):
-            compressed_file = open(full_path, 'r')
-            decompressed_content = zlib.decompress(compressed_file.read())
-            compressed_file.close()
-            cached_file.write(decompressed_content)
-        else:
-            raise FileNotFoundInRepository(file_name)
-
-    def _retrieve_raw_file(self, file_name, cached_file):
-        """ Retrieves the file directly from the source """
-        full_path = self._make_file_uri(file_name)
-        if os.path.exists(full_path):
-            raw_file = open(full_path, 'rb')
-            cached_file.write(raw_file.read())
-            raw_file.close()
-        else:
-            raise FileNotFoundInRepository(file_name)
-
-
-class RemoteFetcher(Fetcher):
-    """ Retrieves files from the local cache if found, and from
-    remote otherwise
-    """
-
-    def __init__(self, repo_url, cache_dir = None):
-        super(RemoteFetcher, self).__init__(repo_url, cache_dir)
-        self._user_agent      = cvmfs.__package_name__ + "/" + cvmfs.__version__
-        self._default_headers = { 'User-Agent': self._user_agent }
-
-    def _download_content_and_store(self, cached_file, file_url):
-        response = requests.get(file_url, stream=True,
-                                          headers=self._default_headers)
-        if response.status_code != requests.codes.ok:
-            raise FileNotFoundInRepository(file_url)
-        for chunk in response.iter_content(chunk_size=4096):
-            if chunk:
-                cached_file.write(chunk)
-
-    def _download_content_and_decompress(self, cached_file, file_url):
-        response = requests.get(file_url, stream=False,
-                                          headers=self._default_headers)
-        if response.status_code != requests.codes.ok:
-            raise FileNotFoundInRepository(file_url)
-        decompressed_content = zlib.decompress(response.content)
-        cached_file.write(decompressed_content)
-
-    def _retrieve_file(self, file_name, cached_file):
-        file_url = self._make_file_uri(file_name)
-        self._download_content_and_decompress(cached_file, file_url)
-
-    def _retrieve_raw_file(self, file_name, cached_file):
-        file_url = self._make_file_uri(file_name)
-        self._download_content_and_store(cached_file, file_url)
 
 
 class Repository(object):
